Use subquery to retrieve a single content per task properly
The subquery returns a single content.id per task, by grouping contents
by task and then taking the minimum content.id for each task.
We can then use a normal query which makes use of this subquery to
retrieve the actual data we need (the full content and task).
Since it is a subquery the database only returns anything once we call
the outer query by looping over the contents.